python/requirements.txt
torch
transformers
sentence-transformers
mysql-connector-python
requests
